+++ demo/foundrydemo/settings.py
@@ -11,8 +11,8 @@ from foundry import settings as foundry_settings
 FOUNDRY = {
     'has_javascript': True,
     'has_ajax': True,
-    'sms_gateway_api_key': 'FCF2F73F-382C-47EA-930E-262F4A08B7C2',
-    'sms_gateway_password': 'ca2fbf32',
+    'sms_gateway_api_key': '',
+    'sms_gateway_password': '',
     'layers': ('basic',)
 }
 
@@ -27,10 +27,10 @@ TEMPLATE_DEBUG = DEBUG
 
 DATABASES = {
     'default': {
-        'ENGINE': 'sqlite3', # Add 'postgresql_psycopg2', 'postgresql', 'mysql', 'sqlite3' or 'oracle'.
-        'NAME': '/tmp/foundrydemo.db', # Or path to database file if using sqlite3.
-        'USER': '', # Not used with sqlite3.
-        'PASSWORD': '', # Not used with sqlite3.
+        'ENGINE': 'mysql', # Add 'postgresql_psycopg2', 'postgresql', 'mysql', 'sqlite3' or 'oracle'.
+        'NAME': 'foundrydemo', # Or path to database file if using sqlite3.
+        'USER': 'foundrydemo', # Not used with sqlite3.
+        'PASSWORD': 'foundrydemo', # Not used with sqlite3.
         'HOST': '', # Set to empty string for localhost. Not used with sqlite3.
         'PORT': '', # Set to empty string for default. Not used with sqlite3.
     }
@@ -116,6 +116,7 @@ ROOT_URLCONF = 'foundrydemo.urls'
 
 INSTALLED_APPS = (
     'foundrydemo',
+    'downloads',
     'friends',
     'foundry',
     'section',
@@ -151,6 +152,8 @@ INSTALLED_APPS = (
     'compressor', # custom - add to paster
     'jmbo_analytics', # add to paster
     'analytics', # add to paster
+    'gunicorn',
+    'sites_groups',
     'django.contrib.auth',
     'django.contrib.comments',
     'django.contrib.contenttypes',