+++ foundry/admin.py
@@ -5,6 +5,7 @@ from django.conf import settings
 from django.contrib.contenttypes.models import ContentType
 
 from preferences.admin import PreferencesAdmin
+from sites_groups.widgets import SitesGroupsWidget
 from jmbo.models import ModelBase
 from jmbo.admin import ModelBaseAdmin
 
@@ -32,7 +33,7 @@ class LinkAdminForm(forms.ModelForm):
     class Meta:
         model = Link
         fields = (
-            'title', 'view_name', 'category', 'target', 'url'
+            'title', 'subtitle', 'view_name', 'category', 'target', 'url'
         )
 
 
@@ -72,7 +73,7 @@ class LinkAdminForm(forms.ModelForm):
 
 class LinkAdmin(admin.ModelAdmin):
     form = LinkAdminForm
-    list_display = ('title', '_get_absolute_url')
+    list_display = ('title', 'subtitle', '_get_absolute_url')
 
     def _get_absolute_url(self, obj):
         url = obj.get_absolute_url()
@@ -85,7 +86,15 @@ class MenuLinkPositionInline(admin.StackedInline):
     model = MenuLinkPosition
 
 
+class MenuAdminForm(forms.ModelForm):
+
+    class Meta:
+        model = Menu
+        widgets = {'sites': SitesGroupsWidget}
+
+
 class MenuAdmin(admin.ModelAdmin):
+    form = MenuAdminForm
     prepopulated_fields = {'slug': ('title',)}
     inlines = [MenuLinkPositionInline]
     list_display = ('title', 'subtitle')
@@ -95,7 +104,15 @@ class NavbarLinkPositionInline(admin.StackedInline):
     model = NavbarLinkPosition
 
 
+class NavbarAdminForm(forms.ModelForm):
+
+    class Meta:
+        model = Navbar
+        widgets = {'sites': SitesGroupsWidget}
+
+
 class NavbarAdmin(admin.ModelAdmin):
+    form = NavbarAdminForm
     prepopulated_fields = {'slug': ('title',)}
     inlines = [NavbarLinkPositionInline]
     list_display = ('title', 'subtitle')
@@ -109,6 +126,7 @@ class ListingAdminForm(forms.ModelForm):
             'title', 'slug', 'subtitle', 'content_type', 'category', 'content', 
             'style', 'count', 'items_per_page', 'sites'
         )       
+        widgets = {'sites': SitesGroupsWidget}
 
     def __init__(self, *args, **kwargs):
         super(ListingAdminForm, self).__init__(*args, **kwargs)
@@ -235,7 +253,15 @@ class PageViewInline(admin.StackedInline):
     model = PageView
 
 
+class PageAdminForm(forms.ModelForm):
+
+    class Meta:
+        model = Page
+        widgets = {'sites': SitesGroupsWidget}
+
+
 class PageAdmin(admin.ModelAdmin):
+    form = PageAdminForm
     list_display = ('title', 'subtitle', 'slug', 'is_homepage')
     prepopulated_fields = {'slug': ('title',)}
     inlines = (PageViewInline,)