+++ foundry/monkey.py
@@ -120,13 +120,19 @@ class LayerAwareSizes(dict):
     
     def get(self, key):
         result = None
+
+        # Handle legacy LAYER marker
         if key.endswith('_LAYER'):
-            prefix = key.replace('LAYER', '') 
-            for layer in settings.FOUNDRY['layers']:
-                result = super(LayerAwareSizes, self).get(prefix + layer)
-                if result is not None:
-                    break
-        else:
+            key = key.replace('_LAYER', '') 
+
+        # Iterate over layers
+        for layer in settings.FOUNDRY['layers']:
+            result = super(LayerAwareSizes, self).get(key + '_' + layer)
+            if result is not None:
+                break
+
+        # Fall back to default
+        if result is None:
             result = super(LayerAwareSizes, self).get(key)
 
         return result
@@ -197,3 +203,23 @@ def BlockNode_render(self, context):
     return result
 
 #BlockNode.render = BlockNode_render
+
+"""FileSystemStorage must be able to handle missing directories. If a foundry 
+based product has a layer 'foo' then collectstatic must not break."""
+import os
+
+from django.core.files.storage import FileSystemStorage
+
+def listdir(self, path):
+        if not self.exists(path):
+            return [], []
+        path = self.path(path)
+        directories, files = [], []
+        for entry in os.listdir(path):
+            if os.path.isdir(os.path.join(path, entry)):
+                directories.append(entry)
+            else:
+                files.append(entry)
+        return directories, files
+
+FileSystemStorage.listdir = listdir